add reward cards
